# Conflicts:
#	app/accessors/TreasureHuntAccessor.java
#	app/controllers/DestinationController.java
#	app/models/Destination.java
#	app/models/DestinationModificationRequest.java
#	app/models/TravellerType.java
#	test/controllers/DestinationControllerTest.java

#story[479]

+++ app/controllers/DestinationController.java
@@ -9,23 +9,18 @@ import factories.DestinationFactory;
 import factories.TravellerTypeFactory;
 import factories.UserFactory;
 import formdata.DestinationFormData;
-import io.ebean.DuplicateKeyException;
 import models.*;
 
 
 import models.commands.Destinations.*;
+import models.commands.General.CommandPage;
 import models.commands.Photos.DeletePhotoCommand;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import play.api.http.MediaRange;
-import play.api.mvc.Request;
 import play.data.DynamicForm;
 import play.data.Form;
 import play.data.FormFactory;
-import play.i18n.Lang;
 import play.libs.Json;
-import play.libs.typedmap.TypedKey;
-import play.libs.typedmap.TypedMap;
 import play.mvc.Controller;
 import play.mvc.Http;
 import play.mvc.Result;
@@ -35,14 +30,14 @@ import views.html.users.destination.*;
 
 import javax.inject.Inject;
 import java.io.File;
-import java.io.IOException;
-import java.security.cert.X509Certificate;
 import java.util.*;
 
 
 import utilities.UtilityFunctions;
 
-
+/**
+ * A controller class for handing destination actions..
+ */
 public class DestinationController extends Controller {
 
     @Inject
@@ -114,16 +109,15 @@ public class DestinationController extends Controller {
      */
     public Result indexDestination(Http.Request request) {
         User user = User.getCurrentUser(request);
-        DestinationFactory destFactory = new DestinationFactory();
 
         if (user != null) {
-            user.getCommandManager().setAllowedType(DestinationPageCommand.class);
+            user.getCommandManager().setAllowedPage(CommandPage.DESTINATION);
 
             CountryUtils.updateCountries();
 
             List<Destination> destinations = user.getDestinations();
 
-            List<Destination> allDestinations = Destination.find.all();
+            List<Destination> allDestinations = Destination.find().all();
 
             return ok(indexDestination.render(destinations, allDestinations, destFactory, user));
 
@@ -153,7 +147,7 @@ public class DestinationController extends Controller {
             }
         }
 
-        user.getCommandManager().setAllowedType(DestinationPageCommand.class);
+//        user.getCommandManager().setAllowedPage(DestinationPageCommand.class);
 
         boolean inEditMode = false;
 
@@ -514,6 +508,7 @@ public class DestinationController extends Controller {
 
                 return redirect(routes.DestinationController.indexDestination());
             }
+
         } else {
             return redirect(routes.UserController.userindex());
         }
@@ -525,12 +520,16 @@ public class DestinationController extends Controller {
 
     /** Perform form validation for the edit/create destination form
      *  Returns a request containing the form with errors if errors found,
-     *  null if no errors found */
+     *  null if no errors found
+     * @param destId the id of the given destination
+     * @param user the id of the user
+     * @param request the http request
+     * @return validates the destination form and checks for errors and if there are matching destinations,
+     * flashes messages if any are found.
+     */
     private Result validateEditCreateForm(Http.Request request, User user, Integer destId) {
         Form<DestinationFormData> destForm;
         destForm = formFactory.form(DestinationFormData.class).bindFromRequest(request);
-
-//        System.out.println(destForm);
 
         // check if private and public destinations already exist.
         // Cannon use .get() on form unless there are no errors
@@ -577,22 +576,21 @@ public class DestinationController extends Controller {
     }
 
 
-
     /**
      * This method handles when an admin rejects a destination modification request
      * the modification request is deleted and the admin is redirected to the index
      * admin page.
      *
-     * @param request
+     * @param request the http request
      * @param destModReqId the id of the destination modification request under review
      * @return given proper authorisation redirect to index admin page
      */
     public Result destinationModificationReject(Http.Request request, Integer destModReqId) {
         User currentUser = User.getCurrentUser(request);
         if (currentUser != null) {
-            Admin currentAdmin = Admin.find.query().where().eq("userId", currentUser.userid).findOne();
+            Admin currentAdmin = Admin.find().query().where().eq("userId", currentUser.getUserid()).findOne();
             if (currentAdmin != null) {
-                DestinationModificationRequest modReq = DestinationModificationRequest.find.query().where().eq("id", destModReqId).findOne();
+                DestinationModificationRequest modReq = DestinationModificationRequest.find().query().where().eq("id", destModReqId).findOne();
                 if (modReq != null) {
 
                     modReq.delete();
@@ -614,16 +612,16 @@ public class DestinationController extends Controller {
      * are used to update the destination, then the modification request is
      * is deleted and the admin is redirected to the index admin page.
      *
-     * @param request
+     * @param request the http request
      * @param destModReqId the id of the destination modification request under review
      * @return given proper authorisation redirect to index admin page
      */
     public Result destinationModificationAccept(Http.Request request, Integer destModReqId) {
         User currentUser = User.getCurrentUser(request);
         if (currentUser != null) {
-            Admin currentAdmin = Admin.find.query().where().eq("userId", currentUser.userid).findOne();
+            Admin currentAdmin = Admin.find().query().where().eq("userId", currentUser.getUserid()).findOne();
             if (currentAdmin != null) {
-                DestinationModificationRequest modReq = DestinationModificationRequest.find.query().where().eq("id", destModReqId).findOne();
+                DestinationModificationRequest modReq = DestinationModificationRequest.find().query().where().eq("id", destModReqId).findOne();
                 if (modReq != null) {
 
                     Destination oldDestination = modReq.getOldDestination();
@@ -676,8 +674,8 @@ public class DestinationController extends Controller {
             JsonNode node = request.body().asJson().get("photoid");
             String photoid = node.textValue();
             photoid = photoid.replace("\"", "");
-            UserPhoto photo = UserPhoto.find.byId(Integer.parseInt(photoid));
-            Destination destination = Destination.find.byId(destId);
+            UserPhoto photo = UserPhoto.find().byId(Integer.parseInt(photoid));
+            Destination destination = Destination.find().byId(destId);
             if (destination != null || photo != null) {
                 if (photo.getUser().getUserid() == user.getUserid()) {
                     //add checks for private destinations here once destinations have been merged in.
@@ -703,9 +701,8 @@ public class DestinationController extends Controller {
         return ok();
     }
 
-
     public Result unlinkAndDelete(Http.Request request, int photoId){
-        UserPhoto photo = UserPhoto.find.byId(photoId);
+        UserPhoto photo = UserPhoto.find().byId(photoId);
         DeletePhotoCommand deletePhotoCommand = new DeletePhotoCommand(photo);
         User user = User.getCurrentUser(request);
         if(user != null){
@@ -724,18 +721,18 @@ public class DestinationController extends Controller {
 
     /**
      * Unlinks the UserPhoto from any destinations and then deletes it
-     * @param request
-     * @param photoId
-     * @return
+     * @param request The HTTP request
+     * @param photoId The Id of the photo being unlinked
+     * @return success if the unlinking was successful.
      */
     public Result unlinkPhotoFromDestinationAndDelete(Http.Request request, int photoId) {
-        UserPhoto photo = UserPhoto.find.byId(photoId);
-        if (photo != null) {
-            for (Destination destination : photo.getDestinations()) {
-                unlinkPhotoFromDestination(request, photoId, destination.getDestId());
+        UserPhoto photo = UserPhoto.find().byId(photoId);
+            if (photo != null) {
+                for (Destination destination : photo.getDestinations()) {
+                    unlinkPhotoFromDestination(request, photoId, destination.getDestId());
+                }
+                UserPhoto.deletePhoto(photoId);
             }
-            UserPhoto.deletePhoto(photoId);
-        }
 
         return ok();
 
@@ -752,20 +749,18 @@ public class DestinationController extends Controller {
      *         badRequest if the destination and photo were not linked     *
      */
     public Result unlinkPhotoFromDestination(Http.Request request, int photoId, int destId) {
-
         User user = User.getCurrentUser(request);
-        UserPhoto photo = UserPhoto.find.byId(photoId);
-        Destination destination = Destination.find.byId(destId);
+        UserPhoto photo = UserPhoto.find().byId(photoId);
+        Destination destination = Destination.find().byId(destId);
 
         if (user == null) return redirect(routes.UserController.userindex());
         if (photo == null) return notFound("No photo found with that id");
         if (destination == null) return notFound("No destination found with that id");
         // This block checks if the user is the owner of either the photo or the destination.
         // If not the owner then returns an unauthorized error else proceeds as usual.
-        if (destination.getUser().getUserid() != user.getUserid()) {
-            if (photo.getUser().getUserid() != user.getUserid()) {
-                return unauthorized("You cannot unlink this photo from this destination as neither of those belong to you.");
-            }
+        if (destination.getUser().getUserid() != user.getUserid()
+                && photo.getUser().getUserid() != user.getUserid()) {
+            return unauthorized("You cannot unlink this photo from this destination as neither of those belong to you.");
         }
         if (!photo.getDestinations().contains(destination))
             return badRequest("The destination was not linked to this photo");
@@ -786,7 +781,7 @@ public class DestinationController extends Controller {
     public Result getTravellerTypes(Http.Request request, Integer destId) {
         User user = User.getCurrentUser(request);
         if (user != null) {
-            return ok(Json.toJson(Destination.find.byId(destId).travellerTypes));
+            return ok(Json.toJson(Destination.find().byId(destId).getTravellerTypes()));
         } else {
             return redirect(routes.UserController.userindex());
         }
@@ -802,14 +797,14 @@ public class DestinationController extends Controller {
     public Result getPhotos(Http.Request request, Integer destId) {
         User user = User.getCurrentUser(request);
         if(user != null){
-            Destination destination = Destination.find.byId(destId);
+            Destination destination = Destination.find().byId(destId);
             List<UserPhoto> photos;
             if(destination.getIsPublic() && !user.userIsAdmin()) {
-                photos = Destination.find.byId(destId).getUserPhotos();
+                photos = Destination.find().byId(destId).getUserPhotos();
                 DestinationFactory destinationFactory = new DestinationFactory();
                 destinationFactory.removePrivatePhotos(photos, user.getUserid());
             } else {
-                photos = Destination.find.byId(destId).getUserPhotos();
+                photos = Destination.find().byId(destId).getUserPhotos();
             }
 
             return ok(Json.toJson(photos));
@@ -822,12 +817,13 @@ public class DestinationController extends Controller {
      * Returns a photo file based on a photo with a given photo id
      *
      * @param request the HTTP request
+     * @param photoId the id for a given photo
      * @return the photo file
      */
     public Result getPhoto(Http.Request request, Integer photoId) {
         User user = User.getCurrentUser(request);
         if(user != null){
-            UserPhoto photo = UserPhoto.find.byId(photoId);
+            UserPhoto photo = UserPhoto.find().byId(photoId);
             if (photo.getUser().getUserid() == user.getUserid() || photo.isPublic() || user.userIsAdmin()) {
                 return ok(Json.toJson(photo));
             } else {
@@ -848,7 +844,7 @@ public class DestinationController extends Controller {
     public Result getDestination(Http.Request request, Integer destId){
         User user = User.getCurrentUser(request);
         if (user != null) {
-            Destination destination = Destination.find.byId(destId);
+            Destination destination = Destination.find().byId(destId);
             if (destination.getIsPublic() || destination.getUser().getUserid() == user.getUserid() || user.userIsAdmin()) {
                 return ok(Json.toJson(destination));
             } else {
@@ -867,8 +863,8 @@ public class DestinationController extends Controller {
      * @return the destination as a json
      */
     public Result getDestinationOwner(Http.Request request, Integer destId) {
-        Destination destination = Destination.find.byId(destId);
-        User user = User.find.query().where().eq("userid", destination.getUser().getUserid()).findOne();
+        Destination destination = Destination.find().byId(destId);
+        User user = User.find().query().where().eq("userid", destination.getUser().getUserid()).findOne();
         if (user != null) {
             return ok(Json.toJson(user.getUserid()));
         } else {
@@ -970,19 +966,17 @@ public class DestinationController extends Controller {
     }
 
 
-
-
     /**
      * Returns an image file to the requester, accepts the UserPhoto id to send back the correct image.
      * @param request the photo
-     * @param destId
-     * @return
+     * @param destId the id of the destination tht is having it's primary photo set
+     * @return success if setting primary photo was successful, not found if destination or photo not found, unauthorized otherwise.
      */
     public Result servePrimaryPicture(Http.Request request, Integer destId) {
         if (destId != null) {
             UserPhoto primaryPicture = DestinationFactory.getPrimaryPicture(destId);
             if (primaryPicture != null) {
-                System.out.println("Sending image back");
+                logger.debug("Sending image back");
                 return ok(new File(primaryPicture.getUrlWithPath()));
             } else {
                 //should be 404 but then console logs an error
@@ -1002,6 +996,5 @@ public class DestinationController extends Controller {
 //        return ok(googlePlacesMapDocumentationExample.render(user));
         return ok(googlePlacesMapDocumentationExample.render(user));
     }
-
 
 }
